requirements.txt
browsermob-proxy
django
django-celery
ipython
selenium
